Add logic to handle frozen player votes
Freezing votes work by disabling the ability to vote or unvote for players who have a casted vote when the ability triggers. A new function: freeze_player_votes is added to update a list of frozen players. is_valid_vote and is_valid_unvote have been updated to return early if the player is in the frozen player list.

Conditions:

1. Only staff members may freeze votes.
1. Staff members votes cannot be frozen.
1. No logic to unfreeze votes at the moment, at this mechanic should last until EoD.