Check no_lynch against player values
Previously, to check if no_lynch was on the vote_rights, we tested wether no_lynch was in the player columns. Not sure why, it was always returning False, and thus, no_lynch appended no matter If it was already present on vote_config.csv. This check has been fixed by comparing against the column values, which is a list.