Fix persistent majority reached flag
If the previous day ends by a player reaching absolute majority, the flag will not be reset upon a new day, and thus, no new votes will be counted (nor any game action).

This fix checks If the last vote count ID is older than the day start post and If the majority_reached flag is set to True. In this case,  it sets it back to False.