Add a command to set up a lylo
LyLo is a situation where, in some games,
votes cannot be removed once casted. To allow
for such a behaviour, this commit adds a GM
command available for both the game master and
the moderators: lylo

Lylo sets a boolean to True in the vcount instance,
which in turns disables the unvote action by returning
False when is_valid_unvote is triggered.